requirements.txt
requests
flask
langgraph
langchain
openai
textblob
gtts
python-dotenv
fpdf
streamlit
python-docx
pydub
